Move create_case utils to module that uses them
These differ from the ones that are used more frequently in the form processor tests. A future test will move those to where these used to be.
