+++ corehq/motech/tests/test_value_source.py
@@ -1,6 +1,7 @@
 import doctest
 from datetime import datetime, timedelta
 from uuid import uuid4
+from unittest.mock import patch
 
 from django.test import SimpleTestCase, TestCase
 
@@ -8,10 +9,15 @@ import attr
 from schema import Use
 
 import corehq.motech.value_source
-from corehq.form_processor.models import CommCareCaseIndex, CommCareCase
+from corehq.form_processor.backends.sql.processor import FormProcessorSQL
+from corehq.form_processor.interfaces.processor import ProcessedForms
+from corehq.form_processor.models import (
+    CaseTransaction,
+    CommCareCase,
+    CommCareCaseIndex,
+    XFormInstance,
+)
 from corehq.form_processor.tests.utils import (
-    create_case,
-    create_case_with_index,
     delete_all_xforms_and_cases,
     sharded,
 )
@@ -644,3 +650,32 @@ class TestSupercaseValueSourceSetExternalValue(TestCase):
 def test_doctests():
     results = doctest.testmod(corehq.motech.value_source, optionflags=doctest.ELLIPSIS)
     assert results.failed == 0
+
+
+def create_case(case) -> CommCareCase:
+    form = XFormInstance(
+        form_id=uuid4().hex,
+        xmlns='http://commcarehq.org/formdesigner/form-processor',
+        received_on=case.server_modified_on,
+        user_id=case.owner_id,
+        domain=case.domain,
+    )
+    transaction = CaseTransaction(
+        type=CaseTransaction.TYPE_FORM,
+        form_id=form.form_id,
+        case=case,
+        server_date=case.server_modified_on,
+    )
+    with patch.object(FormProcessorSQL, "publish_changes_to_kafka"):
+        case.track_create(transaction)
+        processed_forms = ProcessedForms(form, [])
+        FormProcessorSQL.save_processed_models(processed_forms, [case])
+    return CommCareCase.objects.get_case(case.case_id)
+
+
+def create_case_with_index(case, index) -> CommCareCase:
+    case = create_case(case)
+    index.case = case
+    case.track_create(index)
+    case.save(with_tracked_models=True)
+    return case
